add test case for notification detail retrieve

+++ src/main/java/com/odde/massivemailer/service/EmailService.java
@@ -1,6 +1,5 @@
 package com.odde.massivemailer.service;
 
-import com.odde.massivemailer.model.Mail;
 import com.odde.massivemailer.model.Notification;
 
 import java.util.List;
@@ -12,15 +11,13 @@ public interface EmailService {
 
 	List<Notification> getOpenedEmailCountList();
 
-
-
 	void destroyAll();
 
     void setOpenedEmailCountList(int i);
 
-	String  getEmailCounterJson(int email_id);
+	String  getEmailCounterJson(Long email_id);
 
 	void addEmail(int i, String subject);
 
-    void increaseCounterOfEmailByOne(int i, String s);
+    void increaseCounterOfEmailByOne(Long i, String s);
 }
